requirements.txt
#Core
flask==3.0.0
beautifulsoup4==4.12.2
requests==2.31.0
pymongo==4.6.0
dnspython==2.4.2
python-whois==0.8.0

# Machine Learning
tensorflow==2.15.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2

# Flask Extensions
flask-limiter==3.5.0
flask-cors==4.0.0
flasgger==0.9.7.1

# tensorflow-metal==1.1.0
